feat : ajouter de la pop-up + suppression du bouton supprimer pour le theme par défaut ref #81

+++ front-end/src/app/footer-creer-theme/footer-creer-theme.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, Output, EventEmitter, Input} from '@angular/core';
+import {Component, OnInit, AfterViewInit, Output, EventEmitter, Input} from '@angular/core';
 import { Router } from '@angular/router';
 import {FormThemeService} from "../services/formTheme.service";
 import {Theme} from "../../models/theme.models";
@@ -12,7 +12,7 @@ import {PatientService} from "../services/patient.service";
   templateUrl: './footer-creer-theme.component.html',
   styleUrls: ['./footer-creer-theme.component.scss']
 })
-export class FooterCreerThemeComponent {
+export class FooterCreerThemeComponent implements OnInit,AfterViewInit{
   // @ts-ignore
   nom : string ="";
   // @ts-ignore
@@ -34,16 +34,7 @@ export class FooterCreerThemeComponent {
     this.formThemeService.imageSubject$.subscribe(images => {
       this.images = images;
     });
-    const bouton = document.getElementById("boutonChangeant") as HTMLButtonElement;
-    if(this.theme == undefined){
-      this.bas_jaune = true;
-      this.supprimer = false;
-      bouton.innerHTML = "Partager la Création";
-    }else{
-      this.bas_jaune = false;
-      this.supprimer = true;
-      bouton.innerHTML = "Supprimer";
-    }
+
   }
   retourListeTheme() {
     this.themeService.setEditTheme(undefined);
@@ -78,16 +69,32 @@ export class FooterCreerThemeComponent {
 
   clickPartagerOuSupprimer(){
     if(this.supprimer) {
-      this.themeService.removeTheme(this.theme);
-      this.themeService.setEditTheme(undefined);
-      const patient = this.patientService.patientSelectionne$;
-      let patientSelect: Patient = this.patientService.getPatientById(patient.value?.id as number);
-      patientSelect.setThemes(this.themeService.listeThemes$.value);
+      let bool = confirm('Êtes-vous sûr de vouloir supprimer ce thème ?');
+      if(bool) {
+        this.themeService.removeTheme(this.theme);
+        this.themeService.setEditTheme(undefined);
+        const patient = this.patientService.patientSelectionne$;
+        let patientSelect: Patient = this.patientService.getPatientById(patient.value?.id as number);
+        patientSelect.setThemes(this.themeService.listeThemes$.value);
 
-      this.router.navigate(['/liste-theme']);
+        this.router.navigate(['/liste-theme']);
+      }
     }
     if(this.bas_jaune){
       this.router.navigate(['/partager-theme']);
+    }
+  }
+
+  ngAfterViewInit(): void {
+    const bouton = document.getElementById("boutonChangeant") as HTMLButtonElement;
+    if(this.theme == undefined){
+      this.bas_jaune = true;
+      this.supprimer = false;
+      bouton.innerHTML = "Partager la Création";
+    }else{
+      this.bas_jaune = false;
+      this.supprimer = true;
+      bouton.innerHTML = "Supprimer";
     }
   }
 }
